+++ test/private/unescapeJsdoc.test.js
@@ -13,14 +13,7 @@ module.exports = (tests) => {
     }
   );
 
-  tests.add('`unescapeJsdoc` unescapes one multiline comment end.', () => {
-    strictEqual(unescapeJsdoc('*\\/'), '*/');
+  tests.add('`unescapeJsdoc` unescapes multiline comment ends.', () => {
+    strictEqual(unescapeJsdoc('*/ *\\/ *\\\\/ *\\\\\\/'), '*/ */ *\\/ *\\\\/');
   });
-
-  tests.add(
-    '`unescapeJsdoc` unescapes multiple multiline comment ends.',
-    () => {
-      strictEqual(unescapeJsdoc('*\\/ *\\/'), '*/ */');
-    }
-  );
 };